requirements.txt
solana==0.30.2
pyTelegramBotAPI==4.14.0
requests==2.31.0
pytest
base58==2.1.1
python-dotenv==1.0.0
solders==0.18.1
spl-token==0.2.0
anchorpy==0.18.0
construct==2.10.68
httpx==0.23.3
